requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
av==15.1.0
avro==1.12.0
beautifulsoup4==4.14.2
billiard==4.2.2
boto3==1.40.21
botocore==1.40.61
Brotli==1.1.0
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coloredlogs==15.0.1
compressed-rtf==1.0.7
constantly==23.10.4
crispy-tailwind==1.0.3
cron_descriptor==2.0.5
cryptography==43.0.3
cssselect2==0.8.0
daphne==4.2.1
dataclasses-json==0.6.7
distro==1.9.0
Django==5.2.5
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-csp==4.0
django-filter==25.1
django-markdown-deux==1.0.6
django-simpleui==2025.6.24
django-temporalio==1.4.0
django-timezone-field==7.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
docopt==0.6.2
ebcdic==1.1.1
elementpath==4.8.0
elevenlabs==2.13.0
extract-msg==0.28.7
fastavro==1.12.0
fonttools==4.60.1
frozenlist==1.8.0
greenlet==3.2.4
gTTS==2.5.4
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
humanfriendly==10.0
humanize==4.13.0
hyperlink==21.0.0
idna==3.11
IMAPClient==2.1.0
incremental==24.7.2
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.4
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-openai==0.3.32
langchain-text-splitters==0.3.11
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.4
langsmith==0.4.21
lxml==6.0.1
Markdown==3.9
markdown2==2.5.4
marshmallow==3.26.1
msgpack==1.1.2
multidict==6.7.0
mypy_extensions==1.1.0
nexus-rpc==1.1.0
num2words==0.5.14
numpy==2.3.2
olefile==0.47
openai==1.102.0
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pydyf==0.11.0
pyOpenSSL==24.2.1
pyphen==0.17.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
redis==7.0.1
regex==2025.8.29
requests==2.32.5
requests-toolbelt==1.0.0
s3transfer==0.13.1
scikit-learn==1.7.1
scipy==1.16.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SpeechRecognition==3.8.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
temporalio==1.18.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
types-protobuf==6.32.1.20250918
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.5.0
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
wikipedia==1.4.0
xmlschema==2.5.1
xxhash==3.5.0
yarl==1.22.0
zope.interface==8.0.1
zopfli==0.2.3.post1
zstandard==0.24.0
